requirements.txt
asgiref==3.2.3
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-allauth==0.41.0
django-debug-toolbar==2.2
django-filter==2.2.0
django-import-export==2.1.0
django-model-utils==4.0.0
django-multiselectfield==0.1.12
django-rest-auth==0.9.5
djangorestframework==3.11.0
et-xmlfile==1.0.1
idna==2.9
jdcal==1.4.1
Markdown==3.2.1
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
tablib==1.1.0
titlecase==0.12.0
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
